// src/app/entity-content/barcle/barcle.content.types.ts
import { RGBA } from 'ngx-color';

export interface BarcleContent {
    isEmitted: boolean;
    amplitudes: Uint8Array;
    startColor: RGBA;
    endColor: RGBA;
    multiplier: number;
    sampleCount: number;
    shadowBlur: number;
    randomizeColors: boolean;
    baseRadius: number;
    ringSize: number;
    fillCenter: boolean;
}